+++ CHANGELOG.md
@@ -7,20 +7,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+
+### Changed
+- Replace manual JS dependency resource handling with `opensavvy` plugin.
+- Replace `useContextRecievers` with `useContextParameters`.
+- Submodule readme generated for all projects, not only for root
+- Replace `wasm` plugin configuration block with `wasmJs`
+
+### Deprecated
+- `useContextRecievers`
+- `wasm` configuration block
+
+### Removed
+- `fullStackApplication` configuration. Replaced by optional field in `fullStack`
+- Jupyter integration
+
+### Fixed
+
+### Security
+
+## 0.17.x
+
+### Added
+
 - kotlinx-io dependency in version catalog
 
 ### Changed
+
 - Use the new jvm executable option for full-stack instead of gradle application plugin.
 
-### Deprecated
-
 ### Removed
+
 - `application` option
 
 ### Fixed
-- Fix readme generation
 
-### Security
+- Fix readme generation
 
 ## 0.16.0-kotlin-2.1.0 - 2025-01-02